+++ src/steamship/agents/tools/audio_transcription/fetch_audio_urls_from_rss_tool.py
@@ -1,11 +1,11 @@
 import re
-from typing import List
+from typing import Any, List, Union
 
 import requests
 
-from steamship import Block
-from steamship.agents.agent_context import AgentContext
-from steamship.tools.tool import Tool, ToolOutput
+from steamship import Block, Task
+from steamship.agents.context import AgentContext
+from steamship.agents.tool import Tool
 from steamship.utils.repl import ToolREPL
 
 
@@ -26,13 +26,13 @@ class FetchAudioUrlsFromRssTool(Tool):
             urls.append(match.group(1))
         return urls
 
-    def run(self, tool_input: List[Block], context: AgentContext) -> ToolOutput:
+    def run(self, tool_input: List[Block], context: AgentContext) -> Union[List[Block], Task[Any]]:
         blocks = []
         for input_block in tool_input:
             if not input_block.is_text():
                 continue
             url = input_block.text
-            urls = self._get_audio_urls(url, context)[:2]
+            urls = self._get_audio_urls(url, context)
             blocks.extend([Block(text=url) for url in urls])
         return blocks
