+++ src/steamship/agents/tools/classification/sentiment_analysis_tool.py
@@ -1,6 +1,6 @@
 from typing import List, Optional
 
-from steamship.tools.text_generation.text_rewriting_tool import TextRewritingTool
+from steamship.agents.tools.text_generation.text_rewrite_tool import TextRewritingTool
 from steamship.utils.repl import ToolREPL
 
 DEFAULT_LABELS = ["positive", "neutral", "negative"]
