+++ src/steamship/agents/tools/classification/zero_shot_classifier_tool.py
@@ -1,6 +1,6 @@
 from typing import List, Optional
 
-from steamship.tools.text_generation.text_rewriting_tool import TextRewritingTool
+from steamship.agents.tools.text_generation.text_rewrite_tool import TextRewritingTool
 from steamship.utils.repl import ToolREPL
 
 DEFAULT_LABELS = ["question", "complaint", "suggestion"]
